requirements.txt
streamlit
google-generativeai
google-genai
PyPDF2
reportlab
scikit-learn
numpy
